+++ app/models/event.rb
@@ -25,6 +25,20 @@ class Event < ActiveRecord::Base
 
   # Model Methods
 
+  # Shows events IF
+  # Project has Blind module disabled
+  # OR Event not set as Only Blinded
+  # OR User is Project Owner
+  # OR User is Unblinded Project Member
+  # OR User is Unblinded Site Member
+  def self.blinding_scope(user)
+    joins(:project)
+      .joins("LEFT OUTER JOIN project_users ON project_users.project_id = projects.id and project_users.user_id = #{user.id}")
+      .joins("LEFT OUTER JOIN site_users ON site_users.project_id = projects.id and site_users.user_id = #{user.id}")
+      .where('projects.blinding_enabled = ? or events.only_unblinded = ? or projects.user_id = ? or project_users.unblinded = ? or site_users.unblinded = ?', false, false, user.id, true, true)
+      .distinct
+  end
+
   def to_param
     slug.blank? ? id : slug
   end