+++ app/models/user.rb
@@ -126,6 +126,14 @@ class User < ActiveRecord::Base
     Design.current.with_project(all_viewable_and_site_projects.select(:id)).blinding_scope(self)
   end
 
+  def all_events
+    Event.current.where(project_id: all_projects.select(:id)).blinding_scope(self)
+  end
+
+  def all_viewable_events
+    Event.current.where(project_id: all_viewable_and_site_projects.select(:id)).blinding_scope(self)
+  end
+
   def all_variables
     Variable.current.with_project(all_projects.pluck(:id))
   end
@@ -136,12 +144,20 @@ class User < ActiveRecord::Base
 
   # Project Editors and Site Editors on that site can modify sheet
   def all_sheets
-    Sheet.current.with_site(all_editable_sites.select(:id)).where(design_id: all_viewable_designs.select(:id))
+    Sheet.current
+         .with_site(all_editable_sites.select(:id))
+         .where(design_id: all_viewable_designs.select(:id))
+         .joins('LEFT OUTER JOIN subject_events ON subject_events.id = sheets.subject_event_id').distinct
+         .where('sheets.subject_event_id IS NULL or (sheets.subject_event_id = subject_events.id and (subject_events.event_id IN (?) or subject_events.event_id IS NULL))', all_viewable_events.select(:id))
   end
 
   # Project Editors and Viewers and Site Members can view sheets
   def all_viewable_sheets
-    Sheet.current.with_site(all_viewable_sites.select(:id)).where(design_id: all_viewable_designs.select(:id))
+    Sheet.current
+         .with_site(all_viewable_sites.select(:id))
+         .where(design_id: all_viewable_designs.select(:id))
+         .joins('LEFT OUTER JOIN subject_events ON subject_events.id = sheets.subject_event_id').distinct
+         .where('sheets.subject_event_id IS NULL or (sheets.subject_event_id = subject_events.id and (subject_events.event_id IN (?) or subject_events.event_id IS NULL))', all_viewable_events.select(:id))
   end
 
   # Only Project Editors or Project Owner can modify randomization