+++ db/migrate/20160222165811_add_only_unblinded_to_events.rb
@@ -1,0 +1,5 @@
+class AddOnlyUnblindedToEvents < ActiveRecord::Migration
+  def change
+    add_column :events, :only_unblinded, :boolean, null: false, default: false
+  end
+end